Undo Engine/Plugins/Runtime/Harmonix/... changelist 32181276 - Changes to file length requirements and quantization on midi file import seems to be causing some bugs. Rolling back for now.
[CL 32498496 by buzz burrowes in 5.4 branch]

+++ Engine/Plugins/Runtime/Harmonix/Source/HarmonixMetasound/Public/HarmonixMetasound/Components/MusicClockComponent.h
@@ -250,7 +250,7 @@ public:
 	float GetSongLengthBeats() const;
 
 	UFUNCTION(BlueprintPure, Category = "Song Data")
-	float GetSongLengthBars() const;
+	int32 GetSongLengthBars() const;
 
 	const FSongMaps& GetSongMaps() const;
